New annotations (#130)
* Fake test annotations (#129)

* add automatic_mallory

* fix tests and logs

* add fake notations

---------


* another fake annotation

* implement multiple parts of a variable tracking, add partial support for unions

* Multiple anotation group (#131)

* add refactor in order to extend easily the program, add support for groups in LLVM

* add change of structure to implement groups

* add refactor

* change in order to compile

* add END to instrumentation

* groups feature finaly made

* add working example in prod

* leader ultimate implementation

* Leader abstraction finally made

* leader abstraction in llvm pass made

* llvm pass

* add targets_types

* add working example leader

---------


* Feature nodes count (#132)

* add node parametizer

* fix docker count

* Change to english and delete comments

* delete comments and change to english

---------


* Add parameter to keep dots, time limit and delete electionTallyTest

* add notations

* fixes in state function

* add notations

* add block to process multiple variables

* fix annotations

* Change Candidate state instrumentation

* Add OVERRADE transition name (#134)

* add overrade function name

* Improve Dockerfile

* improve dockerfile

* add new annotations

---------


* add new anottation

* Optimizations (#135)

* add overrade function on SUT

* add optimizations

---------


---------

+++ src/commands/export_graphs.rs
@@ -25,6 +25,9 @@ pub struct ExportGraphsArgs {
 
     #[arg(short = 'f', long = "format", value_enum, default_value = "png")]
     pub format: OutputFormat,
+
+    #[arg(long = "keep-dot", default_value = "false")]
+    pub keep_dot: bool,
 }
 
 pub fn run(args: ExportGraphsArgs, logger: &Logger, sh: &Shell) -> Result<()> {
@@ -94,8 +97,9 @@ pub fn run(args: ExportGraphsArgs, logger: &Logger, sh: &Shell) -> Result<()> {
                     .with_context(|| {
                         format!("graphviz 'dot' failed generating {}", out_path.display())
                     })?;
-                logger.debug("Cleaning up temporary DOT file");
-                let _ = fs::remove_file(&tmp_dot);
+                if !args.keep_dot {
+                    let _ = fs::remove_file(&tmp_dot);
+                }
                 logger.success(format!("Wrote {}", out_path.display()));
             }
         }
@@ -128,6 +132,7 @@ mod tests {
             log_path: "tests/export_graphs_test/simple_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -153,6 +158,7 @@ mod tests {
             log_path: "tests/export_graphs_test/multi_node_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -183,6 +189,7 @@ mod tests {
             log_path: "tests/export_graphs_test/simple_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Png,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -218,6 +225,7 @@ mod tests {
             log_path: "tests/export_graphs_test/simple_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Pdf,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -253,6 +261,7 @@ mod tests {
             log_path: "tests/export_graphs_test/simple_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -273,6 +282,7 @@ mod tests {
             log_path: "tests/export_graphs_test/empty_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -296,6 +306,7 @@ mod tests {
             log_path: "tests/export_graphs_test/nonexistent.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -315,6 +326,7 @@ mod tests {
             log_path: "tests/export_graphs_test/complex_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
@@ -350,6 +362,7 @@ mod tests {
             log_path: "tests/export_graphs_test/simple_events.log".to_string(),
             output_dir: output_dir.to_str().unwrap().to_string(),
             format: OutputFormat::Dot,
+            keep_dot: false,
         };
 
         let logger = create_test_logger();
